requirements.txt
streamlit
open-clip-torch
pillow
pymupdf
scikit-learn
torch
tqdm
matplotlib
fpdf
